CW: Routes: Fixed the querying to the database of the booking function and added a total price and tickets prices querying and variable to the function, created query and route to retrieve screening times for movies
Form: Added the additional form inputs needed for booking and fixed the variable names in the search form

Booking: Rearranged and added loops to show a list of the movies and screenings and
added javascript to use AGAX to use the retieve screening route to get the screenings of only the selected movie
Fixed labels and choices showing multiple times

Search: Fixed the page so it functions off of the navbar need to implement navbar icon to access the page,
working on a movie-detail page that will show the movie contents in the search results and can be used to build the movie pages.
Also made it so search results appear as cards and has the movie title that can be selected and then resirected to movie page (Need to implement)

Stylesheet: Fixed layout of the search and booking to be centre of page and fit webpage style.