CW: Long commit message
Done: Updated the services page to connect to the database
and retrieve the menu items from there and loop them into product cards,
worked on the styling of them a bit as well.

Fixed the input text to be black to be visible and edited the
styling of the login and registration (This one still needs work) to be fit the rest of the website.
The backgrounds are still not showing on these pages.

TO-Do: Still working on the booking system as the screenigns are still
not showing but roberto has made the movie details page that will make
things easier as he has got the screenings to show with filter.
Need to implement using the url variables of movie id and screening id
to automatically fill out booking form.
Make tickets not go to negative.